backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-pptx==0.6.22
PyPDF2==3.0.1
chromadb==0.4.18
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
Pillow==10.1.0
pandas==2.1.3
aiofiles==23.2.1
